--- test/common/utxobased/keymanager/coins/bitcoinsv.spec.ts
@@ -1,145 +1,0 @@
-import { expect } from 'chai'
-import { describe, it } from 'mocha'
-
-import {
-  addressToScriptPubkey,
-  AddressTypeEnum,
-  BIP43PurposeTypeEnum,
-  privateKeyToWIF,
-  pubkeyToScriptPubkey,
-  scriptPubkeyToAddress,
-  ScriptTypeEnum,
-  seedOrMnemonicToXPriv,
-  wifToPrivateKey,
-  xprivToXPub,
-  xpubToPubkey,
-} from '../../../../../src/common/utxobased/keymanager/keymanager'
-import { NetworkEnum } from '../../../../../src/common/plugin/types'
-
-describe('bitcoin sv mnemonic to xprv test vectors as compared with iancoleman', () => {
-  const mnemonic =
-    'abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about'
-  it('bip44 mnemonic to xpriv mainnet', () => {
-    const resultLegacy = seedOrMnemonicToXPriv({
-      seed: mnemonic,
-      network: NetworkEnum.Mainnet,
-      type: BIP43PurposeTypeEnum.Legacy,
-      coin: 'bitcoinsv',
-    })
-    expect(resultLegacy).to.equal(
-      'xprv9ydzpAw8scxgS53bvJyqSwDvfxDQZZtaJV98SYjZto3Pg7MCsPBjCcYqUtnWPRNayEXUcSYZDvXux545bHZwda7YUWvReJiRkx38VXathgK'
-    )
-  })
-
-  it('bip44 mnemonic to xpriv testnet', () => {
-    const resultLegacyTestnet = seedOrMnemonicToXPriv({
-      seed: mnemonic,
-      network: NetworkEnum.Testnet,
-      type: BIP43PurposeTypeEnum.Legacy,
-      coin: 'bitcoinsv',
-    })
-    expect(resultLegacyTestnet).to.equal(
-      'tprv8fPDJN9UQqg6pFsQsrVxTwHZmXLvHpfGGcsCA9rtnatUgVtBKxhtFeqiyaYKSWydunKpjhvgJf6PwTwgirwuCbFq8YKgpQiaVJf3JCrNmkR'
-    )
-  })
-})
-
-describe('bitcoin sv bip32 prefix tests for the conversion from xpriv to xpub', () => {
-  it('bip44 xpriv to xpub mainnet', () => {
-    const resultLegacy = xprivToXPub({
-      xpriv:
-        'xprv9ydzpAw8scxgS53bvJyqSwDvfxDQZZtaJV98SYjZto3Pg7MCsPBjCcYqUtnWPRNayEXUcSYZDvXux545bHZwda7YUWvReJiRkx38VXathgK',
-      network: NetworkEnum.Mainnet,
-      type: BIP43PurposeTypeEnum.Legacy,
-      coin: 'bitcoinsv',
-    })
-    expect(resultLegacy).to.equals(
-      'xpub6CdMDgU2hzWyeZ852LWqp5AfDz3ty2cRfi4jEw9BT8aNYugMQvVykQsKLARZdbqKKp7yTviJdL1N9saYLmJNKD1rwVAwLTmU8r8qKeoyG4R'
-    )
-  })
-
-  it('bip44 xpriv to xpub testnet', () => {
-    const resultLegacyTestnet = xprivToXPub({
-      xpriv:
-        'tprv8fPDJN9UQqg6pFsQsrVxTwHZmXLvHpfGGcsCA9rtnatUgVtBKxhtFeqiyaYKSWydunKpjhvgJf6PwTwgirwuCbFq8YKgpQiaVJf3JCrNmkR',
-      network: NetworkEnum.Testnet,
-      type: BIP43PurposeTypeEnum.Legacy,
-      coin: 'bitcoinsv',
-    })
-    expect(resultLegacyTestnet).to.equals(
-      'tpubDC5FSnBiZDMmhiuCmWAYsLwgLYrrT9rAqvTySfuCCrgsWz8wxMXUS9Tb9iVMvcRbvFcAHGkMD5Kx8koh4GquNGNTfohfk7pgjhaPCdXpoba'
-    )
-  })
-})
-
-describe('bitcoin sv xpub to address tests;  generate valid addresses by calling xpubToPubkey, pubkeyToScriptPubkey and scriptPubkeyToAddress', () => {
-  /*
-    These methods were cross verified using ian colemans bip32 website https://iancoleman.io/bip39/
-    using the same seed as in other tests (abandon, ...)
-    */
-
-  it('given an xpub, generate p2pkh address and cross verify script pubkey result', () => {
-    const pubkeyP2PKH = xpubToPubkey({
-      xpub:
-        'xpub6CdMDgU2hzWyeZ852LWqp5AfDz3ty2cRfi4jEw9BT8aNYugMQvVykQsKLARZdbqKKp7yTviJdL1N9saYLmJNKD1rwVAwLTmU8r8qKeoyG4R',
-      network: NetworkEnum.Mainnet,
-      type: BIP43PurposeTypeEnum.Legacy,
-      bip44ChangeIndex: 0,
-      bip44AddressIndex: 0,
-      coin: 'bitcoinsv',
-    })
-    const scriptPubkeyP2PKH = pubkeyToScriptPubkey({
-      pubkey: pubkeyP2PKH,
-      scriptType: ScriptTypeEnum.p2pkh,
-    }).scriptPubkey
-
-    const p2pkhAddress = scriptPubkeyToAddress({
-      scriptPubkey: scriptPubkeyP2PKH,
-      network: NetworkEnum.Mainnet,
-      addressType: AddressTypeEnum.p2pkh,
-      coin: 'bitcoinsv',
-    })
-    expect(p2pkhAddress.address).to.equals('1K6LZdwpKT5XkEZo2T2kW197aMXYbYMc4f')
-    expect(p2pkhAddress.legacyAddress).to.equals(
-      'bitcoincash:qrr8ftywq56qg3che8hy2r2anuj7ysmy2qwc7tfd96'
-    )
-    const scriptPubkeyP2PKHRoundTrip = addressToScriptPubkey({
-      address: '1K6LZdwpKT5XkEZo2T2kW197aMXYbYMc4f',
-      network: NetworkEnum.Mainnet,
-      addressType: AddressTypeEnum.p2pkh,
-      coin: 'bitcoinsv',
-    })
-    expect(scriptPubkeyP2PKHRoundTrip).to.equals(scriptPubkeyP2PKH)
-  })
-})
-
-describe('bitcoin sv from WIF to private key to WIF', () => {
-  it('take a wif private key', () => {
-    const wifKey = 'KxU83MzcLXP1WJtoFJXMDMcN3z5ykAa9xLdFTDY5XpV4e6Zit9BA'
-    const privateKey = wifToPrivateKey({
-      wifKey,
-      network: NetworkEnum.Mainnet,
-      coin: 'bitcoinsv',
-    })
-    const wifKeyRoundTrip = privateKeyToWIF({
-      privateKey: privateKey,
-      network: NetworkEnum.Mainnet,
-      coin: 'bitcoinsv',
-    })
-    expect(wifKey).to.be.equal(wifKeyRoundTrip)
-  })
-})
-
-describe('bitcoin sv guess script pubkeys from address', () => {
-  // these tests are cross verified with bitcoin core
-  it('p2pkh address to scriptPubkey', () => {
-    const scriptPubkey = addressToScriptPubkey({
-      address: '1K6LZdwpKT5XkEZo2T2kW197aMXYbYMc4f',
-      network: NetworkEnum.Mainnet,
-      coin: 'bitcoinsv',
-    })
-    expect(scriptPubkey).to.equal(
-      '76a914c674ac8e0534044717c9ee450d5d9f25e243645088ac'
-    )
-  })
-})
