+++ test/common/utxobased/keymanager/coins/bitcointransactiontest.spec.ts
@@ -1,6 +1,7 @@
 import { expect } from 'chai'
 import { describe, it } from 'mocha'
 
+import { NetworkEnum } from '../../../../../src/common/plugin/types'
 import {
   addressToScriptPubkey,
   AddressTypeEnum,
@@ -18,268 +19,8 @@ import {
   TxInput,
   TxOutput,
   wifToPrivateKey,
-  xprivToPrivateKey,
-  xprivToXPub,
-  xpubToPubkey,
+  xprivToPrivateKey
 } from '../../../../../src/common/utxobased/keymanager/keymanager'
-import { NetworkEnum } from '../../../../../src/common/plugin/types'
-
-describe('bitcoin mnemonic to xprv test vectors as collected from BIP84, BIP49 and some generated cases to test xpub prefix bytes', () => {
-  const mnemonic =
-    'abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about'
-  it('bip84 mnemonic to xpriv mainnet', () => {
-    const resultSegwit = seedOrMnemonicToXPriv({
-      seed: mnemonic,
-      network: NetworkEnum.Mainnet,
-      type: BIP43PurposeTypeEnum.Segwit,
-      coin: 'bitcoin',
-    })
-    expect(resultSegwit).to.equal(
-      'zprvAdG4iTXWBoARxkkzNpNh8r6Qag3irQB8PzEMkAFeTRXxHpbF9z4QgEvBRmfvqWvGp42t42nvgGpNgYSJA9iefm1yYNZKEm7z6qUWCroSQnE'
-    )
-  })
-
-  it('bip84 mnemonic to xpriv testnet', () => {
-    const resultSegwitTestnet = seedOrMnemonicToXPriv({
-      seed: mnemonic,
-      network: NetworkEnum.Testnet,
-      type: BIP43PurposeTypeEnum.Segwit,
-      coin: 'bitcoin',
-    })
-    expect(resultSegwitTestnet).to.equal(
-      'vprv9K7GLAaERuM58PVvbk1sMo7wzVCoPwzZpVXLRBmum93gL5pSqQCAAvZjtmz93nnnYMr9i2FwG2fqrwYLRgJmDDwFjGiamGsbRMJ5Y6siJ8H'
-    )
-  })
-
-  it('bip49 mnemonic to xpriv mainnet', () => {
-    const resultWrappedSegwit = seedOrMnemonicToXPriv({
-      seed: mnemonic,
-      network: NetworkEnum.Mainnet,
-      type: BIP43PurposeTypeEnum.WrappedSegwit,
-      coin: 'bitcoin',
-    })
-    expect(resultWrappedSegwit).to.equal(
-      'yprvAHwhK6RbpuS3dgCYHM5jc2ZvEKd7Bi61u9FVhYMpgMSuZS613T1xxQeKTffhrHY79hZ5PsskBjcc6C2V7DrnsMsNaGDaWev3GLRQRgV7hxF'
-    )
-  })
-
-  it('bip49 mnemonic to xpriv testnet', () => {
-    const resultWrappedSegwitTestnet = seedOrMnemonicToXPriv({
-      seed: mnemonic,
-      network: NetworkEnum.Testnet,
-      type: BIP43PurposeTypeEnum.WrappedSegwit,
-      coin: 'bitcoin',
-    })
-    expect(resultWrappedSegwitTestnet).to.equal(
-      'uprv91G7gZkzehuMVxDJTYE6tLivdF8e4rvzSu1LFfKw3b2Qx1Aj8vpoFnHdfUZ3hmi9jsvPifmZ24RTN2KhwB8BfMLTVqaBReibyaFFcTP1s9n'
-    )
-  })
-
-  it('bip44 mnemonic to xpriv mainnet', () => {
-    const resultLegacy = seedOrMnemonicToXPriv({
-      seed: mnemonic,
-      network: NetworkEnum.Mainnet,
-      type: BIP43PurposeTypeEnum.Legacy,
-      coin: 'bitcoin',
-    })
-    expect(resultLegacy).to.equal(
-      'xprv9xpXFhFpqdQK3TmytPBqXtGSwS3DLjojFhTGht8gwAAii8py5X6pxeBnQ6ehJiyJ6nDjWGJfZ95WxByFXVkDxHXrqu53WCRGypk2ttuqncb'
-    )
-  })
-
-  it('bip44 mnemonic to xpriv testnet', () => {
-    const resultLegacyTestnet = seedOrMnemonicToXPriv({
-      seed: mnemonic,
-      network: NetworkEnum.Testnet,
-      type: BIP43PurposeTypeEnum.Legacy,
-      coin: 'bitcoin',
-    })
-    expect(resultLegacyTestnet).to.equal(
-      'tprv8fPDJN9UQqg6pFsQsrVxTwHZmXLvHpfGGcsCA9rtnatUgVtBKxhtFeqiyaYKSWydunKpjhvgJf6PwTwgirwuCbFq8YKgpQiaVJf3JCrNmkR'
-    )
-  })
-})
-
-describe('bitcoin bip32 prefix tests for the conversion from xpriv to xpub', () => {
-  it('bip84 xpriv to xpub mainnet', () => {
-    const resultSegwit = xprivToXPub({
-      xpriv:
-        'zprvAdG4iTXWBoARxkkzNpNh8r6Qag3irQB8PzEMkAFeTRXxHpbF9z4QgEvBRmfvqWvGp42t42nvgGpNgYSJA9iefm1yYNZKEm7z6qUWCroSQnE',
-      network: NetworkEnum.Mainnet,
-      type: BIP43PurposeTypeEnum.Segwit,
-      coin: 'bitcoin',
-    })
-    expect(resultSegwit).to.equals(
-      'zpub6rFR7y4Q2AijBEqTUquhVz398htDFrtymD9xYYfG1m4wAcvPhXNfE3EfH1r1ADqtfSdVCToUG868RvUUkgDKf31mGDtKsAYz2oz2AGutZYs'
-    )
-  })
-
-  it('bip84 xpriv to xpub mainnet', () => {
-    const resultSegwitTestnet = xprivToXPub({
-      xpriv:
-        'vprv9K7GLAaERuM58PVvbk1sMo7wzVCoPwzZpVXLRBmum93gL5pSqQCAAvZjtmz93nnnYMr9i2FwG2fqrwYLRgJmDDwFjGiamGsbRMJ5Y6siJ8H',
-      network: NetworkEnum.Testnet,
-      type: BIP43PurposeTypeEnum.Segwit,
-      coin: 'bitcoin',
-    })
-    expect(resultSegwitTestnet).to.equals(
-      'vpub5Y6cjg78GGuNLsaPhmYsiw4gYX3HoQiRBiSwDaBXKUafCt9bNwWQiitDk5VZ5BVxYnQdwoTyXSs2JHRPAgjAvtbBrf8ZhDYe2jWAqvZVnsc'
-    )
-  })
-
-  it('bip49 xpriv to xpub mainnet', () => {
-    const resultWrappedSegwit = xprivToXPub({
-      xpriv:
-        'yprvAHwhK6RbpuS3dgCYHM5jc2ZvEKd7Bi61u9FVhYMpgMSuZS613T1xxQeKTffhrHY79hZ5PsskBjcc6C2V7DrnsMsNaGDaWev3GLRQRgV7hxF',
-      network: NetworkEnum.Mainnet,
-      type: BIP43PurposeTypeEnum.WrappedSegwit,
-      coin: 'bitcoin',
-    })
-    expect(resultWrappedSegwit).to.equals(
-      'ypub6Ww3ibxVfGzLrAH1PNcjyAWenMTbbAosGNB6VvmSEgytSER9azLDWCxoJwW7Ke7icmizBMXrzBx9979FfaHxHcrArf3zbeJJJUZPf663zsP'
-    )
-  })
-
-  it('bip49 xpriv to xpub testnet', () => {
-    const resultWrappedSegwitTestnet = xprivToXPub({
-      xpriv:
-        'uprv91G7gZkzehuMVxDJTYE6tLivdF8e4rvzSu1LFfKw3b2Qx1Aj8vpoFnHdfUZ3hmi9jsvPifmZ24RTN2KhwB8BfMLTVqaBReibyaFFcTP1s9n',
-      network: NetworkEnum.Testnet,
-      type: BIP43PurposeTypeEnum.WrappedSegwit,
-      coin: 'bitcoin',
-    })
-    expect(resultWrappedSegwitTestnet).to.equals(
-      'upub5EFU65HtV5TeiSHmZZm7FUffBGy8UKeqp7vw43jYbvZPpoVsgU93oac7Wk3u6moKegAEWtGNF8DehrnHtv21XXEMYRUocHqguyjknFHYfgY'
-    )
-  })
-
-  it('bip44 xpriv to xpub mainnet', () => {
-    const resultLegacy = xprivToXPub({
-      xpriv:
-        'xprv9xpXFhFpqdQK3TmytPBqXtGSwS3DLjojFhTGht8gwAAii8py5X6pxeBnQ6ehJiyJ6nDjWGJfZ95WxByFXVkDxHXrqu53WCRGypk2ttuqncb',
-      network: NetworkEnum.Mainnet,
-      type: BIP43PurposeTypeEnum.Legacy,
-      coin: 'bitcoin',
-    })
-    expect(resultLegacy).to.equals(
-      'xpub6BosfCnifzxcFwrSzQiqu2DBVTshkCXacvNsWGYJVVhhawA7d4R5WSWGFNbi8Aw6ZRc1brxMyWMzG3DSSSSoekkudhUd9yLb6qx39T9nMdj'
-    )
-  })
-
-  it('bip44 xpriv to xpub testnet', () => {
-    const resultLegacyTestnet = xprivToXPub({
-      xpriv:
-        'tprv8fVU32aAEuEPeH1WYx3LhXtSFZTRaFqjbFNPaJZ9R8fCVja44tSaUPZEKGpMK6McUDkWWMvRiVfKR3Wzei6AmLoTNYHMAZ9KtvVTLZZdhvA',
-      network: NetworkEnum.Testnet,
-      type: BIP43PurposeTypeEnum.Legacy,
-      coin: 'bitcoin',
-    })
-    expect(resultLegacyTestnet).to.equals(
-      'tpubDCBWBScQPGv4Xk3JSbhw6wYYpayMjb2eAYyArpbSqQTbLDpphHGAetB6VQgVeftLML8vDSUEWcC2xDi3qJJ3YCDChJDvqVzpgoYSuT52MhJ'
-    )
-  })
-})
-
-describe('bitcoin xpub to address tests;  generate valid addresses by calling xpubToPubkey, pubkeyToScriptPubkey and scriptPubkeyToAddress', () => {
-  /*
-  These methods were cross verified using ian colemans bip32 website https://iancoleman.io/bip39/
-  using the same seed as in other tests (abandon, ...)
-  */
-
-  it('given an xpub, generate p2pkh address and cross verify script pubkey result', () => {
-    const pubkeyP2PKH = xpubToPubkey({
-      xpub:
-        'xpub6BosfCnifzxcFwrSzQiqu2DBVTshkCXacvNsWGYJVVhhawA7d4R5WSWGFNbi8Aw6ZRc1brxMyWMzG3DSSSSoekkudhUd9yLb6qx39T9nMdj',
-      network: NetworkEnum.Mainnet,
-      type: BIP43PurposeTypeEnum.Legacy,
-      bip44ChangeIndex: 0,
-      bip44AddressIndex: 0,
-      coin: 'bitcoin',
-    })
-
-    const scriptPubkeyP2PKH = pubkeyToScriptPubkey({
-      pubkey: pubkeyP2PKH,
-      scriptType: ScriptTypeEnum.p2pkh,
-    }).scriptPubkey
-
-    const p2pkhAddress = scriptPubkeyToAddress({
-      scriptPubkey: scriptPubkeyP2PKH,
-      network: NetworkEnum.Mainnet,
-      addressType: AddressTypeEnum.p2pkh,
-      coin: 'bitcoin',
-    }).address
-    expect(p2pkhAddress).to.equals('1LqBGSKuX5yYUonjxT5qGfpUsXKYYWeabA')
-    const scriptPubkeyP2PKHRoundTrip = addressToScriptPubkey({
-      address: '1LqBGSKuX5yYUonjxT5qGfpUsXKYYWeabA',
-      network: NetworkEnum.Mainnet,
-      addressType: AddressTypeEnum.p2pkh,
-      coin: 'bitcoin',
-    })
-    expect(scriptPubkeyP2PKHRoundTrip).to.equals(scriptPubkeyP2PKH)
-  })
-
-  it('given an ypub, generate p2wpkh-p2sh address and cross verify script pubkey result', () => {
-    const pubkeyP2WPKHP2SH = xpubToPubkey({
-      xpub:
-        'ypub6Ww3ibxVfGzLrAH1PNcjyAWenMTbbAosGNB6VvmSEgytSER9azLDWCxoJwW7Ke7icmizBMXrzBx9979FfaHxHcrArf3zbeJJJUZPf663zsP',
-      network: NetworkEnum.Mainnet,
-      type: BIP43PurposeTypeEnum.WrappedSegwit,
-      bip44ChangeIndex: 0,
-      bip44AddressIndex: 0,
-      coin: 'bitcoin',
-    })
-    const scriptPubkeyP2WPKHP2SH = pubkeyToScriptPubkey({
-      pubkey: pubkeyP2WPKHP2SH,
-      scriptType: ScriptTypeEnum.p2wpkhp2sh,
-    }).scriptPubkey
-    const p2wpkhp2shAddress = scriptPubkeyToAddress({
-      scriptPubkey: scriptPubkeyP2WPKHP2SH,
-      network: NetworkEnum.Mainnet,
-      addressType: AddressTypeEnum.p2sh,
-      coin: 'bitcoin',
-    }).address
-    expect(p2wpkhp2shAddress).to.equals('37VucYSaXLCAsxYyAPfbSi9eh4iEcbShgf')
-    const scriptPubkeyP2WPKHP2SHRoundTrip = addressToScriptPubkey({
-      address: '37VucYSaXLCAsxYyAPfbSi9eh4iEcbShgf',
-      network: NetworkEnum.Mainnet,
-      addressType: AddressTypeEnum.p2sh,
-      coin: 'bitcoin',
-    })
-    expect(scriptPubkeyP2WPKHP2SHRoundTrip).to.equals(scriptPubkeyP2WPKHP2SH)
-  })
-
-  it('given a zpub, generate p2wpkh address and cross verify script pubkey result', () => {
-    const pubkeyP2WPKH = xpubToPubkey({
-      xpub:
-        'zpub6rFR7y4Q2AijBEqTUquhVz398htDFrtymD9xYYfG1m4wAcvPhXNfE3EfH1r1ADqtfSdVCToUG868RvUUkgDKf31mGDtKsAYz2oz2AGutZYs',
-      network: NetworkEnum.Mainnet,
-      type: BIP43PurposeTypeEnum.Segwit,
-      bip44ChangeIndex: 0,
-      bip44AddressIndex: 0,
-      coin: 'bitcoin',
-    })
-    const scriptPubkeyP2WPKH = pubkeyToScriptPubkey({
-      pubkey: pubkeyP2WPKH,
-      scriptType: ScriptTypeEnum.p2wpkh,
-    }).scriptPubkey
-    const p2wpkhAddress = scriptPubkeyToAddress({
-      scriptPubkey: scriptPubkeyP2WPKH,
-      network: NetworkEnum.Mainnet,
-      addressType: AddressTypeEnum.p2wpkh,
-      coin: 'bitcoin',
-    }).address
-    expect(p2wpkhAddress).to.equals(
-      'bc1qcr8te4kr609gcawutmrza0j4xv80jy8z306fyu'
-    )
-    const scriptPubkeyP2WPKHRoundTrip = addressToScriptPubkey({
-      address: 'bc1qcr8te4kr609gcawutmrza0j4xv80jy8z306fyu',
-      network: NetworkEnum.Mainnet,
-      addressType: AddressTypeEnum.p2wpkh,
-      coin: 'bitcoin',
-    })
-    expect(scriptPubkeyP2WPKHRoundTrip).to.be.equal(scriptPubkeyP2WPKH)
-  })
-})
 
 describe('bitcoin bip32 seed, aka airbitz seed, to xpriv. Taken from official bip32 test vectors', () => {
   it('seed to xpriv', () => {
@@ -288,7 +29,7 @@ describe('bitcoin bip32 seed, aka airbitz seed, to xpriv. Taken from official bi
         '//z59vPw7ern5OHe29jV0s/MycbDwL26t7SxrquopaKfnJmWk5CNioeEgX57eHVyb2xpZmNgXVpXVFFOS0hFQg==',
       network: NetworkEnum.Mainnet,
       type: BIP43PurposeTypeEnum.Legacy,
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
     expect(result).to.equal(
       'xprv9vHkqa6EV4sPZHYqZznhT2NPtPCjKuDKGY38FBWLvgaDx45zo9WQRUT3dKYnjwih2yJD9mkrocEZXo1ex8G81dwSM1fwqWpWkeS3v86pgKt'
@@ -302,7 +43,7 @@ describe('bitcoin from bip32 seed to private key', () => {
       seed: 'Ha/lqCZZCz+VWN2eGrHYuG/ag6S8w67rldgfCrlcPWI=',
       network: NetworkEnum.Mainnet,
       type: BIP43PurposeTypeEnum.Legacy,
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
     const privateKey = xprivToPrivateKey({
       xpriv,
@@ -310,33 +51,16 @@ describe('bitcoin from bip32 seed to private key', () => {
       type: BIP43PurposeTypeEnum.Legacy,
       bip44ChangeIndex: 0,
       bip44AddressIndex: 0,
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
     const wifKey = privateKeyToWIF({
       privateKey: privateKey,
       network: NetworkEnum.Mainnet,
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
     expect(wifKey).to.equal(
       'L4XwV2KLfsWYzaHbZRBGfrYBabYVCRUz8N6juyUDbMxYBPAeVdug'
     )
-  })
-})
-
-describe('bitcoin from WIF to private key buffer to WIF', () => {
-  it('take a wif private key', () => {
-    const wifKey = 'L2uPYXe17xSTqbCjZvL2DsyXPCbXspvcu5mHLDYUgzdUbZGSKrSr'
-    const privateKey = wifToPrivateKey({
-      wifKey,
-      network: NetworkEnum.Mainnet,
-      coin: 'bitcoin',
-    })
-    const wifKeyRoundTrip = privateKeyToWIF({
-      privateKey: privateKey,
-      network: NetworkEnum.Mainnet,
-      coin: 'bitcoin',
-    })
-    expect(wifKey).to.be.equal(wifKeyRoundTrip)
   })
 })
 
@@ -347,7 +71,7 @@ describe('bitcoin get script pubkeys from address', () => {
       address: '1KRMKfeZcmosxALVYESdPNez1AP1mEtywp',
       network: NetworkEnum.Mainnet,
       addressType: AddressTypeEnum.p2pkh,
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
     expect(scriptPubkey).to.equal(
       '76a914ca0d36044e0dc08a22724efa6f6a07b0ec4c79aa88ac'
@@ -358,7 +82,7 @@ describe('bitcoin get script pubkeys from address', () => {
       address: 'bc1qcr8te4kr609gcawutmrza0j4xv80jy8z306fyu',
       network: NetworkEnum.Mainnet,
       addressType: AddressTypeEnum.p2wpkh,
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
     expect(scriptPubkey).to.equal(
       '0014c0cebcd6c3d3ca8c75dc5ec62ebe55330ef910e2'
@@ -369,7 +93,7 @@ describe('bitcoin get script pubkeys from address', () => {
       address: '2Mu9hifsg4foPLkyo9i1isPWTobnNmXL3Qk',
       network: NetworkEnum.Testnet,
       addressType: AddressTypeEnum.p2sh,
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
     expect(scriptPubkey).to.equal(
       'a91414e4e7810e5120cc68d55d03b36cf66a9eadc27087'
@@ -380,51 +104,7 @@ describe('bitcoin get script pubkeys from address', () => {
       address: 'tb1qrp33g0q5c5txsp9arysrx4k6zdkfs4nce4xj0gdcccefvpysxf3q0sl5k7',
       network: NetworkEnum.Testnet,
       addressType: AddressTypeEnum.p2wsh,
-      coin: 'bitcoin',
-    })
-    expect(scriptPubkey).to.equal(
-      '00201863143c14c5166804bd19203356da136c985678cd4d27a1b8c6329604903262'
-    )
-  })
-})
-
-describe('bitcoin guess script pubkeys from address', () => {
-  // these tests are cross verified with bitcoin core
-  it('p2pkh address to scriptPubkey', () => {
-    const scriptPubkey = addressToScriptPubkey({
-      address: '1KRMKfeZcmosxALVYESdPNez1AP1mEtywp',
-      network: NetworkEnum.Mainnet,
-      coin: 'bitcoin',
-    })
-    expect(scriptPubkey).to.equal(
-      '76a914ca0d36044e0dc08a22724efa6f6a07b0ec4c79aa88ac'
-    )
-  })
-  it('p2wpkh address to scriptPubkey', () => {
-    const scriptPubkey = addressToScriptPubkey({
-      address: 'bc1qcr8te4kr609gcawutmrza0j4xv80jy8z306fyu',
-      network: NetworkEnum.Mainnet,
-      coin: 'bitcoin',
-    })
-    expect(scriptPubkey).to.equal(
-      '0014c0cebcd6c3d3ca8c75dc5ec62ebe55330ef910e2'
-    )
-  })
-  it('p2sh testnet address to scriptPubkey', () => {
-    const scriptPubkey = addressToScriptPubkey({
-      address: '2Mu9hifsg4foPLkyo9i1isPWTobnNmXL3Qk',
-      network: NetworkEnum.Testnet,
-      coin: 'bitcoin',
-    })
-    expect(scriptPubkey).to.equal(
-      'a91414e4e7810e5120cc68d55d03b36cf66a9eadc27087'
-    )
-  })
-  it('p2wsh mainnet address to scriptPubkey', () => {
-    const scriptPubkey = addressToScriptPubkey({
-      address: 'tb1qrp33g0q5c5txsp9arysrx4k6zdkfs4nce4xj0gdcccefvpysxf3q0sl5k7',
-      network: NetworkEnum.Testnet,
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
     expect(scriptPubkey).to.equal(
       '00201863143c14c5166804bd19203356da136c985678cd4d27a1b8c6329604903262'
@@ -438,7 +118,7 @@ describe('bitcoin address to electrum script hash', () => {
       address: '1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa',
       network: NetworkEnum.Mainnet,
       addressType: AddressTypeEnum.p2pkh,
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
     expect(scriptPubkey).to.equal(
       '76a91462e907b15cbf27d5425399ebf6f0fb50ebb88f1888ac'
@@ -456,41 +136,41 @@ describe('bitcoin transaction creation and signing test', () => {
   const privateKey = wifToPrivateKey({
     wifKey,
     network: NetworkEnum.Mainnet,
-    coin: 'bitcoin',
+    coin: 'bitcoin'
   })
   const scriptPubkey: string = pubkeyToScriptPubkey({
     pubkey: privateKeyToPubkey(privateKey),
-    scriptType: ScriptTypeEnum.p2pkh,
+    scriptType: ScriptTypeEnum.p2pkh
   }).scriptPubkey
   const segwitScriptPubkey: string = pubkeyToScriptPubkey({
     pubkey: privateKeyToPubkey(privateKey),
-    scriptType: ScriptTypeEnum.p2wpkh,
+    scriptType: ScriptTypeEnum.p2wpkh
   }).scriptPubkey
   const wrappedSegwitScriptPubkey: string = pubkeyToScriptPubkey({
     pubkey: privateKeyToPubkey(privateKey),
-    scriptType: ScriptTypeEnum.p2wpkhp2sh,
+    scriptType: ScriptTypeEnum.p2wpkhp2sh
   }).scriptPubkey
   const wrappedSegwitRedeemScript: string | undefined = pubkeyToScriptPubkey({
     pubkey: privateKeyToPubkey(privateKey),
-    scriptType: ScriptTypeEnum.p2wpkhp2sh,
+    scriptType: ScriptTypeEnum.p2wpkhp2sh
   }).redeemScript
   const address: string = scriptPubkeyToAddress({
     scriptPubkey: scriptPubkey,
     network: NetworkEnum.Mainnet,
     coin: 'bitcoin',
-    addressType: AddressTypeEnum.p2pkh,
+    addressType: AddressTypeEnum.p2pkh
   }).address
   const segwitAddress: string = scriptPubkeyToAddress({
     scriptPubkey: segwitScriptPubkey,
     network: NetworkEnum.Mainnet,
     coin: 'bitcoin',
-    addressType: AddressTypeEnum.p2wpkh,
+    addressType: AddressTypeEnum.p2wpkh
   }).address
   const wrappedSegwitAddress: string = scriptPubkeyToAddress({
     scriptPubkey: wrappedSegwitScriptPubkey,
     network: NetworkEnum.Mainnet,
     coin: 'bitcoin',
-    addressType: AddressTypeEnum.p2sh,
+    addressType: AddressTypeEnum.p2sh
   }).legacyAddress
 
   it('Create transaction with one legacy input and one output', async () => {
@@ -524,8 +204,8 @@ describe('bitcoin transaction creation and signing test', () => {
             scriptPubkey +
             // locktime
             '00000000',
-          index: 0,
-        },
+          index: 0
+        }
       ],
       outputs: [
         {
@@ -533,16 +213,16 @@ describe('bitcoin transaction creation and signing test', () => {
             address: '1KRMKfeZcmosxALVYESdPNez1AP1mEtywp',
             network: NetworkEnum.Mainnet,
             addressType: AddressTypeEnum.p2pkh,
-            coin: 'bitcoin',
+            coin: 'bitcoin'
           }),
-          amount: 80000,
-        },
-      ],
+          amount: 80000
+        }
+      ]
     }).psbt
     const hexTxSigned: string = await signTx({
       psbt: base64Tx,
       privateKeys: [privateKey],
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
     expect(hexTxSigned).to.equal(
       '02000000013ebc8203037dda39d482bf41ff3be955996c50d9d4f7cfc3d2097a694a7' +
@@ -563,7 +243,7 @@ describe('bitcoin transaction creation and signing test', () => {
       // prev_tx only for non segwit inputs
       prevTx:
         '0200000001f9f34e95b9d5c8abcd20fc5bd4a825d1517be62f0f775e5f36da944d9452e550000000006b483045022100c86e9a111afc90f64b4904bd609e9eaed80d48ca17c162b1aca0a788ac3526f002207bb79b60d4fc6526329bf18a77135dc5660209e761da46e1c2f1152ec013215801210211755115eabf846720f5cb18f248666fec631e5e1e66009ce3710ceea5b1ad13ffffffff01905f0100000000001976a9148bbc95d2709c71607c60ee3f097c1217482f518d88ac00000000',
-      index: 0,
+      index: 0
     }
 
     const txOutput: TxOutput = {
@@ -571,22 +251,22 @@ describe('bitcoin transaction creation and signing test', () => {
         address: segwitAddress,
         network: NetworkEnum.Mainnet,
         addressType: AddressTypeEnum.p2wpkh,
-        coin: 'bitcoin',
+        coin: 'bitcoin'
       }),
-      amount: 200,
+      amount: 200
     }
 
     const base64Tx: string = createTx({
       inputs: [txInput],
       outputs: Array(nOutputs).fill(txOutput),
       network: NetworkEnum.Mainnet,
-      rbf: false,
+      rbf: false
     }).psbt
 
     const rawtransaction: string = await signTx({
       psbt: base64Tx,
       privateKeys: [privateKey],
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
 
     expect(rawtransaction).to.equal(
@@ -601,7 +281,7 @@ describe('bitcoin transaction creation and signing test', () => {
           '8b26fa4d0238788ffc3a7d96e4169acf6fe993a28791e9e748819ac216ee85b3',
         prevScriptPubkey: segwitScriptPubkey,
         index: i,
-        value: 200,
+        value: 200
       }
     }
 
@@ -609,12 +289,12 @@ describe('bitcoin transaction creation and signing test', () => {
       inputs: txInputs,
       outputs: [txOutput],
       network: NetworkEnum.Mainnet,
-      rbf: false,
+      rbf: false
     }).psbt
     const segwitRawTransaction: string = await signTx({
       psbt: segwitTx,
       privateKeys: Array(nOutputs).fill(privateKey),
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
 
     expect(segwitRawTransaction).to.equal(
@@ -635,7 +315,7 @@ describe('bitcoin transaction creation and signing test', () => {
         '61da46e1c2f1152ec013215801210211755115eabf846720f5cb18f248666fec631e' +
         '5e1e66009ce3710ceea5b1ad13ffffffff01905f0100000000001976a9148bbc95d2' +
         '709c71607c60ee3f097c1217482f518d88ac00000000',
-      index: 0,
+      index: 0
     }
 
     const txInputSegwit: TxInput = {
@@ -644,7 +324,7 @@ describe('bitcoin transaction creation and signing test', () => {
         '8b26fa4d0238788ffc3a7d96e4169acf6fe993a28791e9e748819ac216ee85b3',
       prevScriptPubkey: segwitScriptPubkey,
       index: 1,
-      value: 200,
+      value: 200
     }
 
     const txInputWrappedSegwit: TxInput = {
@@ -654,7 +334,7 @@ describe('bitcoin transaction creation and signing test', () => {
       prevScriptPubkey: wrappedSegwitScriptPubkey,
       redeemScript: wrappedSegwitRedeemScript,
       index: 1,
-      value: 200,
+      value: 200
     }
 
     const txOutputLegacy: TxOutput = {
@@ -662,9 +342,9 @@ describe('bitcoin transaction creation and signing test', () => {
         address: address,
         network: NetworkEnum.Mainnet,
         addressType: AddressTypeEnum.p2pkh,
-        coin: 'bitcoin',
+        coin: 'bitcoin'
       }),
-      amount: 200,
+      amount: 200
     }
 
     const txOutputSegwit: TxOutput = {
@@ -672,9 +352,9 @@ describe('bitcoin transaction creation and signing test', () => {
         address: segwitAddress,
         network: NetworkEnum.Mainnet,
         addressType: AddressTypeEnum.p2wpkh,
-        coin: 'bitcoin',
+        coin: 'bitcoin'
       }),
-      amount: 200,
+      amount: 200
     }
 
     const txOutputWrappedSegwit: TxOutput = {
@@ -682,22 +362,22 @@ describe('bitcoin transaction creation and signing test', () => {
         address: wrappedSegwitAddress,
         network: NetworkEnum.Mainnet,
         addressType: AddressTypeEnum.p2sh,
-        coin: 'bitcoin',
+        coin: 'bitcoin'
       }),
-      amount: 200,
+      amount: 200
     }
 
     const base64Tx: string = createTx({
       inputs: [txInputLegacy, txInputSegwit, txInputWrappedSegwit],
       outputs: [txOutputLegacy, txOutputSegwit, txOutputWrappedSegwit],
       network: NetworkEnum.Mainnet,
-      rbf: false,
+      rbf: false
     }).psbt
 
     const rawtransaction: string = await signTx({
       psbt: base64Tx,
       privateKeys: [privateKey, privateKey, privateKey],
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
     expect(rawtransaction).to.equal(
       '020000000001033ebc8203037dda39d482bf41ff3be955996c50d9d4f7cfc3d2097a694a7b067d000000006b483045022100c10c70dab0e8e88a67bb013d276b88b3d5a52d3cb3ec53aa7d2e7c4704e89dc602207fabdfa3b55b196da7792de9fe436535501fce2fe76439d5361bd5d9dd69bd0901210365db9da3f8a260078a7e8f8b708a1161468fb2323ffda5ec16b261ec1056f455ffffffffb385ee16c29a8148e7e99187a293e96fcf9a16e4967d3afc8f7838024dfa268b0100000000ffffffff9e5d335ec9133a9e5ed426d08c421273594b82ab9876b5beb66716384688f2e901000000171600148bbc95d2709c71607c60ee3f097c1217482f518dffffffff03c8000000000000001976a9148bbc95d2709c71607c60ee3f097c1217482f518d88acc8000000000000001600148bbc95d2709c71607c60ee3f097c1217482f518dc80000000000000017a9142427d83a84e5793ce6f6efc33020d844dd217dbb8700024730440220008a9738e8acdc002ae4bb415ff7368161057ed9b22fd8b671c5acbb47379a69022045ab99fa73730fa367d3972d41b26b4ef337b34c4e593c4f47800ed4ee1c64fd01210365db9da3f8a260078a7e8f8b708a1161468fb2323ffda5ec16b261ec1056f4550247304402201d4f3ee993cfb468dc41387ac618081c36c49ebc9db528155b51d346382c347802201d1c07201f80df3e6e86195444532b6be86e9a429a7eb5a0ae268df9760a974601210365db9da3f8a260078a7e8f8b708a1161468fb2323ffda5ec16b261ec1056f45500000000'
@@ -720,7 +400,7 @@ describe('bitcoin transaction creation and signing test', () => {
         '709c71607c60ee3f097c1217482f518d88ac00000000',
       index: 0,
       // prev_scriptPubkey only relevant for Segwit inputs, but keep mandatory for now before we start handling errors.
-      prevScriptPubkey: scriptPubkey,
+      prevScriptPubkey: scriptPubkey
     }
 
     const txOutput: TxOutput = {
@@ -728,22 +408,22 @@ describe('bitcoin transaction creation and signing test', () => {
         address: address,
         network: NetworkEnum.Mainnet,
         addressType: AddressTypeEnum.p2pkh,
-        coin: 'bitcoin',
+        coin: 'bitcoin'
       }),
-      amount: 200,
+      amount: 200
     }
 
     const base64Tx: string = createTx({
       inputs: [txInput],
       outputs: Array(nOutputs).fill(txOutput),
       network: NetworkEnum.Mainnet,
-      rbf: false,
+      rbf: false
     }).psbt
 
     const hexTxSigned: string = await signTx({
       psbt: base64Tx,
       privateKeys: [privateKey],
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
 
     const txInputs: TxInput[] = Array(nOutputs)
@@ -753,20 +433,20 @@ describe('bitcoin transaction creation and signing test', () => {
         prevTxid:
           '8b26fa4d0238788ffc3a7d96e4169acf6fe993a28791e9e748819ac216ee85b3',
         prevTx: hexTxSigned,
-        index: i,
+        index: i
       }
     }
     const base64TxMulti: string = createTx({
       inputs: txInputs,
       outputs: [txOutput, txOutput],
       network: NetworkEnum.Mainnet,
-      rbf: false,
+      rbf: false
     }).psbt
 
     const hexTxMultiSigned: string = await signTx({
       psbt: base64TxMulti,
       privateKeys: Array(nOutputs).fill(privateKey),
-      coin: 'bitcoin',
+      coin: 'bitcoin'
     })
 
     expect(hexTxMultiSigned).to.equal(
